old-tests/submission/Infraware/Forms/contents/Forms/button_formtarget.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_formTarget</h3>
	</p>
	
	<hr>

	<div id="log"></div> 
	
	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><button id='button_id'>button</button></p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		button.formTarget = "_blank";

		if (typeof(button.formTarget) == "string") {
			test(function() {
				assert_equals(button.formTarget, "_blank", "formTarget attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formTarget attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
